+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/replay/ScaleFactor.kt
@@ -3,12 +3,12 @@ package com.launchdarkly.observability.replay
 import android.view.View
 import kotlin.math.roundToInt
 
-fun calculateScaleFactor(scale: Float?, view: View): Float {
-    if (scale == null) return 1f
-    val density = view.resources.displayMetrics.density
-    return if (density > 0f) scale / density else 1f
+fun calculateScaleFactor(scale: Double?, view: View): Double {
+    if (scale == null) return 1.0
+    val density = view.resources.displayMetrics.density.toDouble()
+    return if (density > 0.0) scale / density else 1.0
 }
 
-fun scaleCoordinate(value: Float, scaleFactor: Float): Int {
+fun scaleCoordinate(value: Float, scaleFactor: Double): Int {
     return (value * scaleFactor).roundToInt()
 }
